# swagger.yml
openapi: 3.0.0
info:
  title: Parent Teacher Portal
  description: A description about the application.
  version: 0.0.1

servers:
  - url: https://localhost:3000
    description: Example app for demonstration locally.

paths:
  /schools:
    post:
      tags:
        - schools
      summary: Add a new school
      description: Add a new school
      operationId: addSchool
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Schools"
        required: true
      responses:
        "200":
          description: Status code indicating success
    put:
      tags:
        - schools
      summary: Update an existing school
      description: Update an existing school by Id
      operationId: updateSchool
      requestBody:
        description: Update an existent school
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Schools"
        required: true
      responses:
        "200":
          description: Status code indicating success
  /schools/{schoolId}:
    get:
      tags:
        - schools
      summary: Get a school
      operationId: getSchoolById
      parameters:
        - name: schoolId
          in: path
          description: ID of school to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schools"
            application/xml:
              schema:
                $ref: "#/components/schemas/Schools"
        "400":
          description: Invalid ID supplied
        "404":
          description: School not found
components:
  schemas:
    Schools:
      required:
        - name
        - location
        - type
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 99
        name:
          type: string
          description: name of the school
          example: University of XYZ
        location:
          type: string
          description: location of the school
          example: 162 Andover Street Sulphur, LA 70663
        type:
          type: string
          description: type of school (public or private)
          enum:
            - public
            - private
          example: private